ISSUES  with vue-loader and copy-webpack-plugin
Dev build works great.

Production build has ISSUES:
1 - The modal templates are dynamically loaded (this work great)
	BUT all of the hard/static coded src links in these templates are not
processed.

2 - I used Copy Copy Webpack Plugin to get the images copied over BUT
that has issues as well. If I copy over all of the images { from:
'asset/img', to: 'asset/img' } images are processed and linked (this
creates duplicates - originals and hashed). THIS IS CURRENTLY THE WAY
THE CONFIG AND COMPONENTS ARE SETUP - IT'S A HACK - BUT IT WORKS.

3 - To solve the duplicates issue I set a subdirectory - 'dynamic-img'
with all of the dynamically modal images to be processed by Copy
Webpack Plugin:

{ from: 'asset/img/dynamic-img', to: 'asset/img/dynamic-img' }

This copies the images as expected and also solved all the duplicate
issues BUT for some reason it breaks the links for the thumbnails. I'm
not sure why!?

Will update when sorted, any help is appreciated.