Solved timing issue with Hero Image:
Since the transitions for the hero image are only
triggered once, in the vue-mount this.sectionIsActive(),
the second half of the transitions are not triggered.
Only the:
  enter | enter-active | enter-to
classes are used.

This in turn means only the enter Javascript
hooks are used. This means any listeners need to
be on the enter hooks:
@before-enter | @enter | @after-enter | @enter-canceled

=======================================================

Need to fix issue with webpack not processing modal-button-images.